+++ src/compas_rv/scene/forceobject.py
@@ -15,6 +15,10 @@ class RhinoForceObject(RhinoDiagramObject):
     edgecolor = ColorDictAttribute(default=Color.blue().darkened(50))
     facecolor = ColorDictAttribute(default=Color.blue().lightened(25))
 
+    vertexgroup = "RV::ForceDiagram::Vertices"
+    edgegroup = "RV::ForceDiagram::Edges"
+    facegroup = "RV::ForceDiagram::Faces"
+
     # =============================================================================
     # Properties
     # =============================================================================