+++ src/compas_rv/scene/formobject.py
@@ -14,6 +14,10 @@ class RhinoFormObject(RhinoDiagramObject):
     edgecolor = ColorDictAttribute(default=Color.green().darkened(50))
     facecolor = ColorDictAttribute(default=Color.green().lightened(25))
 
+    vertexgroup = "RV::FormDiagram::Vertices"
+    edgegroup = "RV::FormDiagram::Edges"
+    facegroup = "RV::FormDiagram::Faces"
+
     # =============================================================================
     # Properties
     # =============================================================================