+++ src/compas_rv/scene/thrustobject.py
@@ -30,6 +30,10 @@ class RhinoThrustObject(RUIMeshObject):
     compressioncolor = ColorAttribute(default=Color.blue())
     tensioncolor = ColorAttribute(default=Color.red())
 
+    vertexgroup = "RV::ThrustDiagram::Vertices"
+    edgegroup = "RV::ThrustDiagram::Edges"
+    facegroup = "RV::ThrustDiagram::Faces"
+
     def __init__(
         self,
         disjoint=True,